Format to JSON
REAL_final is essentially done and formatted the same as our data.json file. Just need to reconfigure marketcap to true values (ie 2000000000 instead of 2B). Exporting to JSON also creates some weirdness in output format. Will need to play around with that some more to get it where it needs to be.